split & mini docs
